Consolidate remaining legacy packet types
