+++ src/main/host/descriptor/socket.c
@@ -13,8 +13,7 @@
 #include "main/host/descriptor/descriptor.h"
 #include "main/host/descriptor/socket.h"
 #include "main/host/descriptor/tcp.h"
-#include "main/host/protocol.h"
-#include "main/routing/packet.h"
+#include "main/network/legacypacket.h"
 #include "main/utility/utility.h"
 
 gsize _legacysocket_getOutputBufferSpaceIncludingTCP(LegacySocket* socket);
