+++ src/main/host/descriptor/tcp.c
@@ -28,8 +28,7 @@
 #include "main/host/descriptor/tcp_cong.h"
 #include "main/host/descriptor/tcp_cong_reno.h"
 #include "main/host/descriptor/tcp_retransmit_tally.h"
-#include "main/host/protocol.h"
-#include "main/routing/packet.h"
+#include "main/network/legacypacket.h"
 #include "main/utility/priority_queue.h"
 #include "main/utility/utility.h"
